light bulb to turn on/off
Click the light bulb to turn on/off with image tag and button the light.
--> learned
Quotes Handling:

Use single quotes ' inside the onclick attribute to avoid conflicts with the double quotes " that wrap the entire attribute value.
Function for Image Toggle:

Added a changeImage function that toggles the image source when the image is clicked.